Gracefully handle unexpected bitbucket payloads
We're getting some payloads where the "commit" element for "commits" are nil.
This makes it so we're a little more explicit with the coercion and more forgiving
in what's handled without error.

Specifically we were get `message undefined` errors on the `commit.message` call within
the commits loop below this change:

```
  commits = (commit_payload.push.changes.flat_map(&:commits) rescue []) || []
  commits.each do |commit|
->  commit.message.scan(/([A-Z]+-[0-9]+(?:-[0-9]+)?)/) do |m|
      m.each do |ref|
        comment_on_record(ref, commit)
      end
```