Show the full name for gitlab repos
In gitlab, you can create groups that namespace the report. So I could have
aha-app/docs and jjbohn/docs. Our current implementation only showed the repo
name, not the group, so it would just show "doc" twice. This actually made
it impossible to tie it to the correct report as well since we use the name
to map the integration.

This fix shows the full name now and fixes both issues.

It is also backwards compatible. To do this, it checks if the saved
repo name is the full name, or just the repo name, then acts accordingly.