allow-numbers-in-github-urls
Dearest Reviewer,

It turns out that github returns a number for the issue id and not
a string. I am not sure what math we will do with the id but when
generating urls we dont care.

Changes:
Adds a to_s to the paths the code builds out
Adds a test
Bumped version number

Becker