DEVELOP-1246: Fix basic auth for v1 Jira integrations
During the Ruby 3 upgrade, we changed how the auth headers were added
in several integrations. The change made to the Jira moved the code
into a method that was being overridden by a method further down in
the class. We need to move the auth code into that method for it to
run.